feat : MusicAdapter 로 Vertical, Horizontal 통합
