+++ android/core/ui/src/main/java/com/ohdodok/catchytape/core/ui/Orientation.kt
@@ -1,0 +1,5 @@
+package com.ohdodok.catchytape.core.ui
+
+enum class Orientation {
+    Horizontal, Vertical
+}
