+++ android/feature/home/src/main/java/com/ohdodok/catchytape/feature/home/HomeFragment.kt
@@ -7,6 +7,8 @@ import androidx.fragment.app.viewModels
 import androidx.navigation.NavDeepLinkRequest
 import androidx.navigation.fragment.findNavController
 import com.ohdodok.catchytape.core.ui.BaseFragment
+import com.ohdodok.catchytape.core.ui.MusicAdapter
+import com.ohdodok.catchytape.core.ui.Orientation
 import com.ohdodok.catchytape.feature.home.databinding.FragmentHomeBinding
 import dagger.hilt.android.AndroidEntryPoint
 
@@ -18,9 +20,7 @@ class HomeFragment : BaseFragment<FragmentHomeBinding>(R.layout.fragment_home) {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         binding.viewModel = viewModel
-        binding.rvRecentlyAddedSong.adapter = MusicHorizontalAdapter()
-
-
+        binding.rvRecentlyAddedSong.adapter = MusicAdapter(Orientation.Horizontal)
 
         binding.ibUpload.setOnClickListener {
             val request = NavDeepLinkRequest.Builder
